Refactor: PEtab code to sub-package
Keep PEtab-related code in a subpackage. An attempt to organize the growing codebase a bit better.
More to be done. Unless anybody is using private functions, this should not break anything.

More parts to be moved separately.
