+++ python/sdist/amici/parameter_mapping.py
@@ -27,7 +27,6 @@ from petab.C import *  # noqa: F403
 
 SingleParameterMapping = Dict[str, Union[numbers.Number, str]]
 SingleScaleMapping = Dict[str, str]
-AmiciModel = Union[amici.Model, amici.ModelPtr]
 
 
 class ParameterMappingForCondition:
@@ -161,184 +160,6 @@ class ParameterMapping(Sequence):
     def free_symbols(self) -> Set[str]:
         """Get IDs of all (symbolic) parameters present in this mapping"""
         return set.union(*(mapping.free_symbols for mapping in self))
-
-
-def fill_in_parameters(
-    edatas: List[amici.ExpData],
-    problem_parameters: Dict[str, numbers.Number],
-    scaled_parameters: bool,
-    parameter_mapping: ParameterMapping,
-    amici_model: AmiciModel,
-) -> None:
-    """Fill fixed and dynamic parameters into the edatas (in-place).
-
-    :param edatas:
-        List of experimental datas :class:`amici.amici.ExpData` with
-        everything except parameters filled.
-    :param problem_parameters:
-        Problem parameters as parameterId=>value dict. Only
-        parameters included here will be set. Remaining parameters will
-        be used as currently set in `amici_model`.
-    :param scaled_parameters:
-        If True, problem_parameters are assumed to be on the scale provided
-        in the parameter mapping. If False, they are assumed
-        to be in linear scale.
-    :param parameter_mapping:
-        Parameter mapping for all conditions.
-    :param amici_model:
-        AMICI model.
-    """
-    if unused_parameters := (
-        set(problem_parameters.keys()) - parameter_mapping.free_symbols
-    ):
-        warnings.warn(
-            "The following problem parameters were not used: "
-            + str(unused_parameters),
-            RuntimeWarning,
-        )
-
-    for edata, mapping_for_condition in zip(edatas, parameter_mapping):
-        fill_in_parameters_for_condition(
-            edata,
-            problem_parameters,
-            scaled_parameters,
-            mapping_for_condition,
-            amici_model,
-        )
-
-
-def fill_in_parameters_for_condition(
-    edata: amici.ExpData,
-    problem_parameters: Dict[str, numbers.Number],
-    scaled_parameters: bool,
-    parameter_mapping: ParameterMappingForCondition,
-    amici_model: AmiciModel,
-) -> None:
-    """Fill fixed and dynamic parameters into the edata for condition
-    (in-place).
-
-    :param edata:
-        Experimental data object to fill parameters into.
-    :param problem_parameters:
-        Problem parameters as parameterId=>value dict. Only
-        parameters included here will be set. Remaining parameters will
-        be used as already set in `amici_model` and `edata`.
-    :param scaled_parameters:
-        If True, problem_parameters are assumed to be on the scale provided
-        in the parameter mapping. If False, they
-        are assumed to be in linear scale.
-    :param parameter_mapping:
-        Parameter mapping for current condition.
-    :param amici_model:
-        AMICI model
-    """
-    map_sim_var = parameter_mapping.map_sim_var
-    scale_map_sim_var = parameter_mapping.scale_map_sim_var
-    map_preeq_fix = parameter_mapping.map_preeq_fix
-    scale_map_preeq_fix = parameter_mapping.scale_map_preeq_fix
-    map_sim_fix = parameter_mapping.map_sim_fix
-    scale_map_sim_fix = parameter_mapping.scale_map_sim_fix
-
-    # Parameter mapping may contain parameter_ids as values, these *must*
-    # be replaced
-
-    def _get_par(model_par, value, mapping):
-        """Replace parameter IDs in mapping dicts by values from
-        problem_parameters where necessary"""
-        if isinstance(value, str):
-            try:
-                # estimated parameter
-                return problem_parameters[value]
-            except KeyError:
-                # condition table overrides must have been handled already,
-                # e.g. by the PEtab parameter mapping, but parameters from
-                # InitialAssignments may still be present.
-                if mapping[value] == model_par:
-                    # prevent infinite recursion
-                    raise
-                return _get_par(value, mapping[value], mapping)
-        if model_par in problem_parameters:
-            # user-provided
-            return problem_parameters[model_par]
-        # prevent nan-propagation in derivative
-        if np.isnan(value):
-            return 0.0
-        # constant value
-        return value
-
-    map_preeq_fix = {
-        key: _get_par(key, val, map_preeq_fix)
-        for key, val in map_preeq_fix.items()
-    }
-    map_sim_fix = {
-        key: _get_par(key, val, map_sim_fix)
-        for key, val in map_sim_fix.items()
-    }
-    map_sim_var = {
-        key: _get_par(key, val, dict(map_sim_fix, **map_sim_var))
-        for key, val in map_sim_var.items()
-    }
-
-    # If necessary, (un)scale parameters
-    if scaled_parameters:
-        unscale_parameters_dict(map_preeq_fix, scale_map_preeq_fix)
-        unscale_parameters_dict(map_sim_fix, scale_map_sim_fix)
-    if not scaled_parameters:
-        # We scale all parameters to the scale they are estimated on, and pass
-        # that information to amici via edata.{parameters,pscale}.
-        # The scaling is necessary to obtain correct derivatives.
-        scale_parameters_dict(map_sim_var, scale_map_sim_var)
-        # We can skip preequilibration parameters, because they are identical
-        # with simulation parameters, and only the latter are used from here
-        # on.
-
-    ##########################################################################
-    # variable parameters and parameter scale
-
-    # parameter list from mapping dict
-    parameters = [
-        map_sim_var[par_id] for par_id in amici_model.getParameterIds()
-    ]
-
-    # scales list from mapping dict
-    scales = [
-        petab_to_amici_scale(scale_map_sim_var[par_id])
-        for par_id in amici_model.getParameterIds()
-    ]
-
-    # plist
-    plist = [
-        ip
-        for ip, par_id in enumerate(amici_model.getParameterIds())
-        if isinstance(parameter_mapping.map_sim_var[par_id], str)
-    ]
-
-    if parameters:
-        edata.parameters = np.asarray(parameters, dtype=float)
-
-    if scales:
-        edata.pscale = amici.parameterScalingFromIntVector(scales)
-
-    if plist:
-        edata.plist = plist
-
-    ##########################################################################
-    # fixed parameters preequilibration
-    if map_preeq_fix:
-        fixed_pars_preeq = [
-            map_preeq_fix[par_id]
-            for par_id in amici_model.getFixedParameterIds()
-        ]
-        edata.fixedParametersPreequilibration = fixed_pars_preeq
-
-    ##########################################################################
-    # fixed parameters simulation
-    if map_sim_fix:
-        fixed_pars_sim = [
-            map_sim_fix[par_id]
-            for par_id in amici_model.getFixedParameterIds()
-        ]
-        edata.fixedParameters = fixed_pars_sim
 
 
 def petab_to_amici_scale(petab_scale: str) -> int:
