+++ python/sdist/amici/petab/__init__.py
@@ -1,0 +1,5 @@
+"""PEtab import related code."""
+
+# ID of model parameter that is to be added to SBML model to indicate
+#  preequilibration
+PREEQ_INDICATOR_ID = "preequilibration_indicator"
